Moved rss management endpoints to channels module
