+++ Tribler/Core/Modules/restapi/my_channel_endpoint.py
@@ -19,7 +19,6 @@ class MyChannelEndpoint(MyChannelBaseEndpoint):
     def __init__(self, session):
         MyChannelBaseEndpoint.__init__(self, session)
         child_handler_dict = {"torrents": MyChannelTorrentsEndpoint,
-                              "rssfeeds": MyChannelRssFeedsEndpoint,
                               "playlists": MyChannelPlaylistsEndpoint,
                               "recheckfeeds": MyChannelRecheckFeedsEndpoint}
         for path, child_cls in child_handler_dict.iteritems():
@@ -268,157 +267,6 @@ class MyChannelModifyTorrentsEndpoint(MyChannelBaseEndpoint):
             return MyChannelBaseEndpoint.return_500(self, request, ex)
 
         return json.dumps({"added": self.torrent_url})
-
-
-class MyChannelRssFeedsEndpoint(MyChannelBaseEndpoint):
-    """
-    This endpoint is responsible for handling requests regarding rss feeds in your channel.
-    """
-
-    def getChild(self, path, request):
-        return MyChannelModifyRssFeedsEndpoint(self.session, path)
-
-    def render_GET(self, request):
-        """
-        .. http:get:: /mychannel/rssfeeds
-
-        Return the RSS feeds in your channel.
-
-            .. sourcecode:: none
-
-                curl -X GET http://localhost:8085/mychannel/rssfeeds
-
-            **Example response**:
-
-            .. sourcecode:: javascript
-
-                {
-                    "rssfeeds": [{
-                        "url": "http://rssprovider.com/feed.xml",
-                    }, ...]
-                }
-        """
-        channel_obj = self.get_my_channel_object()
-        if channel_obj is None:
-            return MyChannelBaseEndpoint.return_404(request)
-
-        rss_list = channel_obj.get_rss_feed_url_list()
-        feeds_list = [{'url': rss_item} for rss_item in rss_list]
-
-        return json.dumps({"rssfeeds": feeds_list})
-
-
-class MyChannelRecheckFeedsEndpoint(MyChannelBaseEndpoint):
-    """
-    This class is responsible for handling requests regarding refreshing rss feeds in your channel.
-    """
-
-    def render_POST(self, request):
-        """
-        .. http:post:: /mychannel/recheckfeeds
-
-        Rechecks all rss feeds in your channel. Returns error 404 if you channel does not exist.
-
-            **Example request**:
-
-            .. sourcecode:: none
-
-                curl -X POST http://localhost:8085/mychannel/recheckrssfeeds
-
-            **Example response**:
-
-            .. sourcecode:: javascript
-
-                {
-                    "rechecked": True
-                }
-
-            :statuscode 404: if you have not created a channel.
-        """
-        channel_obj = self.get_my_channel_object()
-        if channel_obj is None:
-            return MyChannelBaseEndpoint.return_404(request)
-
-        channel_obj.refresh_all_feeds()
-
-        return json.dumps({"rechecked": True})
-
-
-class MyChannelModifyRssFeedsEndpoint(MyChannelBaseEndpoint):
-    """
-    This class is responsible for methods that modify the list of RSS feed URLs (adding/removing feeds).
-    """
-
-    def __init__(self, session, feed_url):
-        MyChannelBaseEndpoint.__init__(self, session)
-        self.feed_url = feed_url
-
-    def render_PUT(self, request):
-        """
-        .. http:put:: /mychannel/rssfeeds/http%3A%2F%2Ftest.com%2Frss.xml
-
-        Add a RSS feed to your channel. Returns error 409 if the supplied RSS feed already exists.
-        Note that the rss feed url should be URL-encoded.
-
-            **Example request**:
-
-            .. sourcecode:: none
-
-                curl -X PUT http://localhost:8085/mychannel/rssfeeds/http%3A%2F%2Ftest.com%2Frss.xml
-
-            **Example response**:
-
-            .. sourcecode:: javascript
-
-                {
-                    "added": True
-                }
-
-            :statuscode 409: (conflict) if the specified RSS URL is already present in your feeds.
-        """
-        channel_obj = self.get_my_channel_object()
-        if channel_obj is None:
-            return MyChannelBaseEndpoint.return_404(request)
-
-        if self.feed_url in channel_obj.get_rss_feed_url_list():
-            request.setResponseCode(http.CONFLICT)
-            return json.dumps({"error": "this rss feed already exists"})
-
-        channel_obj.create_rss_feed(self.feed_url)
-        return json.dumps({"added": True})
-
-    def render_DELETE(self, request):
-        """
-        .. http:delete:: /mychannel/rssfeeds/http%3A%2F%2Ftest.com%2Frss.xml
-
-        Delete a RSS feed from your channel. Returns error 404 if the RSS feed that is being removed does not exist.
-        Note that the rss feed url should be URL-encoded.
-
-            **Example request**:
-
-            .. sourcecode:: none
-
-                curl -X DELETE http://localhost:8085/mychannel/rssfeeds/http%3A%2F%2Ftest.com%2Frss.xml
-
-            **Example response**:
-
-            .. sourcecode:: javascript
-
-                {
-                    "removed": True
-                }
-
-            :statuscode 404: if the specified RSS URL is not in your feed list.
-        """
-        channel_obj = self.get_my_channel_object()
-        if channel_obj is None:
-            return MyChannelBaseEndpoint.return_404(request)
-
-        if self.feed_url not in channel_obj.get_rss_feed_url_list():
-            return MyChannelBaseEndpoint.return_404(request, message="this url is not added to your RSS feeds")
-
-        channel_obj.remove_rss_feed(self.feed_url)
-        return json.dumps({"removed": True})
 
 
 class MyChannelPlaylistsEndpoint(MyChannelBaseEndpoint):
